images/requirements.txt
gitpython==3.1.29
beautifulsoup4==4.11.1
requests==2.28.1